+++ packages/shared/src/index.ts
@@ -2,4 +2,5 @@ export * from "./types";
 export * from "./dates";
 export * from "./kirchspiele";
 export * from "./kirchen-coords";
+export * from "./kategorien";
 export * from "./theme";